+++ android/src/main/java/com/twiliovoicereactnative/VoiceService.java
@@ -191,6 +191,7 @@ public class VoiceService extends Service {
     // accept call
     AcceptOptions acceptOptions = new AcceptOptions.Builder()
       .enableDscp(true)
+      .callMessageListener(new CallMessageListenerProxy())
       .build();
     callRecord.setCall(
       callRecord.getCallInvite().accept(