+++ src/index.tsx
@@ -15,6 +15,8 @@ export { Voice } from './Voice';
 export { AudioDevice } from './AudioDevice';
 export { Call } from './Call';
 export { CallInvite } from './CallInvite';
+export { CallMessage } from './CallMessage';
+export { OutgoingCallMessage } from './OutgoingCallMessage';
 export { CustomParameters } from './type/common';
 export { CallKit } from './type/CallKit';
 export { RTCStats } from './type/RTCStats';