+++ CHANGELOG.md
@@ -3,6 +3,36 @@
 
 Twilio Voice React Native SDK has now reached milestone `beta.5`. Included in this version are the following.
 
+## Features
+
+### Call Message Events (Beta)
+
+- Allow sending and receiving "user-defined" messages during an ongoing Voice Call and during a pending Call Invite.
+- To send a CallMessage, and handle `sent` and `failure` cases:
+```ts
+const message = new CallMessage({
+   content: { key1: 'This is a messsage from the parent call' },
+   contentType: CallMessage.ContentType.ApplicationJson,
+   messageType: CallMessage.MessageType.UserDefinedMessage
+});
+const outgoingCallMessage: OutgoingCallMessage = await call.sendMessage(message);
+
+outgoingCallMessage.addListener(OutgoingCallMessage.Event.Failure, (error) => {
+   // outgoingCallMessage failed, handle error
+});
+
+outgoingCallMessage.addListener(OutgoingCallMessage.Event.Sent, () => {
+    // outgoingCallMessage sent
+});
+```
+- To `receive` a CallMessage:
+```ts
+call.addListener(Call.Event.MessageReceived, (message: CallMessage) => {
+  // callMessage received
+});
+```
+- Related docs: https://www.twilio.com/docs/voice/sdks/call-message-events
+
 ## Fixes
 
 - Fixed and improved the docstrings for the `Voice` and `Call` listeners. The descriptions of the events and listeners should now point to the correct docstrings.
@@ -30,6 +60,23 @@ Twilio Voice React Native SDK has now reached milestone `beta.5`. Included in th
   Please see the `Voice` class API documentation [here](https://github.com/twilio/twilio-voice-react-native/blob/latest/docs/api/voice-react-native-sdk.voice_class.md) for details.
 
   Please see the `CallInvite` class API documentation [here](https://github.com/twilio/twilio-voice-react-native/blob/latest/docs/api/voice-react-native-sdk.callinvite_class.md) for details.
+
+- Call Notifications can be customized on Android.
+  
+  The following features regarding a call notificaiton can now be modified
+  - incoming/outgoing/answered call notification tray icon
+  - name of caller/or recipient
+  
+  The incoming/outgoing/answered call notification tray icon can be changed by adding a drawable resources with the following id to your application
+  - `incoming_call_small_icon` for incoming call notifications
+  - `answered_call_small_icon` for answered call notifications
+  - `outgoing_call_small_icon` for outgoing call notifications
+  
+  The name of the caller/or recipient of a call in the notification can be set by adding the following string resources with the following ids to your application.
+  - `incoming_call_caller_name_text` for incoming call notifications
+  - `outgoing_call_caller_name_text` for outgoing call notifications
+  - `answered_call_caller_name_text` for answered call notifications
+  NOTE: For `incoming_call_caller_name_text` & `answered_call_caller_name_text`, the substring `${from}` will be replaced with the caller and for `outgoing_call_caller_name_text`, the substring `${to}` will be replaced with the recipient of the call (if available, defaulting to "unknown").
 
 ### Platform Specific Changes